refactor: Option Service Facade패턴 적용
